+++ src/main/kotlin/store/controller/StoreController.kt
@@ -1,0 +1,4 @@
+package store.controller
+
+class StoreController {
+}
